Add a test suite to check gobject-introspection data
The test suite acts also as an example for downstream projects to generate
gobject-introspection data. A known-good .gir file is checked in for checking
results in the future. Lastly, a small python uses the gobject-introspection
data to create a small editor window.

The change to gtk/makefile is because scintilla needs to be in a shared library
for the python program. Therefore scintilla.a has to be compiled with -fPIC.